appliance: #183 Phase 7 — strip docker; k3s-only from boot
Operator decision: no migration path for fielded appliances. The
appliance is k3s-only from boot, full re-install via ISO. Phase 7
shifts from "migration tooling" to "remove every trace of docker
from the appliance image".

What's gone
-----------
* ``appliance/mkosi.conf`` — drop ``docker.io``, ``docker-compose``,
  and ``fuse-overlayfs`` packages. (The whole "docker overlay2
  can't nest on overlayfs" workaround is obsolete.)
* ``appliance/mkosi.extra/usr/local/share/spatiumddi/docker-compose
  .yml`` — deleted.
* ``appliance/mkosi.extra/etc/systemd/system/spatium-docker-overlay
  .service`` — deleted; the loop-mounted ``docker-overlay.img`` is
  gone with it.
* ``appliance/mkosi.postinst`` — drop the docker.service + spatium-
  docker-overlay.service enable entries; remove the docker-CLI
  static-binary fetch stanza; tighten the resolved-symlink
  comment.
* ``appliance/mkosi.extra/usr/local/bin/spatium-install`` — drop the
  ``-G sudo,docker`` from the admin useradd, retire the
  "docker volumes" disk-floor comment.
* ``agent/supervisor/spatium_supervisor/docker_api.py`` — deleted.
* ``agent/supervisor/spatium_supervisor/service_lifecycle.py`` (the
  compose path) — deleted.
* ``agent/supervisor/images/supervisor/Dockerfile`` — drop
  ``docker-cli`` + ``docker-cli-compose`` Alpine packages.

What's new / rewritten
----------------------
* ``appliance/mkosi.extra/usr/local/bin/spatiumddi-firstboot`` —
  full rewrite. No dockerd wait, no docker-compose pull, no docker-
  overlay mount-signal. New steps: render a HelmChart CR (with
  ``spec.chartContent`` = base64 of the baked appliance chart tgz
  + ``supervisor.enabled=true``) into
  ``/var/lib/rancher/k3s/server/manifests/spatium-bootstrap.yaml``
  on every boot. k3s's built-in helm-controller picks the CR up
  on first start; the supervisor pod auto-deploys. Readiness wait
  switches from polling ``/health/live`` on docker-bridged 8000
  to polling kubeapi ``/readyz``.
* ``appliance/scripts/bake-images.sh`` — rewritten. Replaces the
  loop-mounted ``docker-overlay.img`` with per-image
  ``.tar.zst`` archives dropped into
  ``appliance/mkosi.extra/var/lib/rancher/k3s/agent/images/``.
  k3s auto-imports anything in that directory at startup
  (containerd's natively-supported air-gap shape) — no firstboot
  shell-out needed. Each image is retagged
  ``ghcr.io/spatiumddi/<name>:${SPATIUMDDI_VERSION}`` before save
  so the chart's image reference resolves locally. Also writes
  ``/usr/lib/spatiumddi/spatiumddi-version`` so firstboot can
  reconcile .env's ``SPATIUMDDI_VERSION`` line to the baked tag.
* ``agent/supervisor/spatium_supervisor/service_lifecycle.py`` —
  was ``service_lifecycle_k3s.py``; renamed + inlines its own
  ``LifecycleResult`` dataclass + ``SUPERVISED_SERVICES`` tuple
  (used to import from the deleted compose-path module).
* ``agent/supervisor/spatium_supervisor/appliance_state.detect_
  runtime()`` — simplified to a constant ``return "k3s"``. The
  pre-Phase-7 systemctl-is-active branch is gone; the function
  survives for callers that still read the runtime tag.
* ``agent/supervisor/spatium_supervisor/heartbeat.py`` — the
  compose-vs-k3s dispatcher (``_select_lifecycle``) is gone;
  imports ``apply_role_assignment`` + ``tear_down_supervised_
  services`` directly from ``service_lifecycle``. Capability flags
  (``can_run_dns_bind9`` / ``can_run_dns_powerdns`` /
  ``can_run_dhcp``) hardcoded to True — every baked appliance
  has every image preloaded.
* ``agent/supervisor/spatium_supervisor/watchdog.py`` — drops
  ``_check_health_compose`` + ``_dispatch_apply`` + the ``_derive_
  status`` (docker-container) helper. Single ``_check_health_k3s``
  path now driven directly by ``check_health``. Auto-heal calls
  ``apply_role_assignment`` (k3s) directly.
* ``agent/supervisor/images/supervisor/entrypoint.sh`` — drops the
  docker.sock gid-juggling that bootstrapped supplementary-group
  membership for the unprivileged spatium user. Adds a
  ``set -a; . /etc/spatiumddi-host/.env; set +a`` step so the
  install-time env (CONTROL_PLANE_URL / BOOTSTRAP_PAIRING_CODE /
  AGENT_GROUP) lands in the supervisor pod's environment without
  a separate Secret or ConfigMap.

Bootstrap flow
--------------
1. Operator boots ISO → spatium-install wizard → reboot.
2. Firstboot writes ``/etc/spatiumddi/.env`` + the bootstrap
   HelmChart manifest.
3. k3s.service starts; helm-controller sees the manifest +
   applies the chart with ``supervisor.enabled=true``.
4. Supervisor pod entrypoint sources ``/etc/spatiumddi-host/.env``
   + starts the Python supervisor; pairs with control plane via
   the BOOTSTRAP_PAIRING_CODE from the wizard.
5. Operator approves the appliance from the Fleet UI; the
   supervisor's heartbeat handler assigns roles → its
   service_lifecycle path PATCHes the HelmChart CR with
   ``dnsBind9.enabled=true`` / etc → helm-controller spins up
   the role pods.

What stayed
-----------
* ``docker-compose.dev.yml`` (control plane dev environment) is
  unchanged. Only the *appliance* drops docker.
* spatium-console's ``docker_ps()`` rewrite to query k3s pods is
  intentionally deferred to a Phase 7 follow-up. The dashboard
  becomes slightly less informative on a Phase-7 appliance until
  then but doesn't break.
