+++ appliance/mkosi.conf
@@ -116,16 +116,19 @@ Packages=
         # directly. Without resolved running, the installed system
         # has no resolver — /etc/resolv.conf is whatever the live
         # system's was (often a stub pointing at ::1 set by
-        # live-config), which dangles after install and
-        # docker-compose pull fails with "lookup registry-1.docker.io
-        # on [::1]:53: connection refused".
+        # live-config), which dangles after install.
         #
         # DNSStubListener=no is set via /etc/systemd/resolved.conf.d/
         # so resolved doesn't bind 127.0.0.53:53 — important because
-        # the appliance's BIND9 container needs the host's :53 free.
+        # k3s's CoreDNS pod needs the host's :53 free for the
+        # in-cluster DNS service.
         systemd-resolved
-        docker.io
-        docker-compose
+        # Issue #183 Phase 7 — docker is gone. The appliance is
+        # k3s-only; the k3s static binary (fetched at build time by
+        # appliance/scripts/fetch-k3s.sh) ships its own embedded
+        # containerd runtime + every CLI we need (kubectl / ctr /
+        # crictl, all symlinks to /usr/local/bin/k3s). No docker.io,
+        # no docker-compose, no docker-cli, no fuse-overlayfs.
         openssh-server
         ca-certificates
         curl
@@ -184,23 +187,14 @@ Packages=
         # logrotate handles /var/log/spatiumddi/firstboot.log so
         # it doesn't grow unbounded across appliance reboots.
         logrotate
-        # zstd decompresses the Phase 4 dev-iteration image tarballs
-        # baked into /usr/local/share/spatiumddi/images/ by
-        # appliance/scripts/bake-images.sh — so the ISO can ship WIP
-        # api/frontend images without depending on a ghcr.io push.
-        # No-op when an ISO is built without baked images.
+        # zstd — required by k3s's air-gap image loader for the
+        # ``*.tar.zst`` tarballs that ship under
+        # ``/var/lib/rancher/k3s/agent/images/`` (the k3s upstream
+        # airgap tarball + our baked SpatiumDDI service images).
+        # k3s imports anything in that dir on startup natively, so
+        # no firstboot shell-out is needed; we just need zstd
+        # available in the rootfs for the embedded importer.
         zstd
-        # fuse-overlayfs — docker's overlay2 storage driver cannot
-        # nest on top of an overlay mount (kernel restriction:
-        # "filesystem ... not supported as upperdir"). The appliance's
-        # /var/lib/docker is itself a kernel-level overlay (baked
-        # image lower + /var/persist upper) so docker can't bring up
-        # its own per-layer mounts on top. fuse-overlayfs is the
-        # user-space FUSE reimplementation — same semantics, no
-        # kernel-nesting restriction — selected via
-        # /etc/docker/daemon.json's storage-driver. ~10-20% slower
-        # than native overlay2 but functional.
-        fuse-overlayfs
         # Phase 4h console dashboard (issue #134). The spatium-console
         # script replaces getty@tty1 on the installed appliance with a
         # Talos-style refresh-on-tick view. python3 ships in the base
